requirements.txt
streamlit
pandas
plotly
PyYAML
requests
